Update to esphome version 2024.2.0
Due to breaking change in esphome version 2024.2.0 the configuration had to be changed. 

In esphome had to be added support for the BME280 sensor over SPI. This change is a breaking change for any existing configurations using the BME280 over I2C as the platform name has been changed from `bme280` to `bme280_i2c`.